fakir: Generate Fake Datasets for Prototyping and Teaching

Create fake datasets that can be used for prototyping and teaching. This package provides a set of functions to generate fake data for a variety of data types, such as dates, addresses, and names. It can be used for prototyping (notably in 'shiny') or as a tool to teach data manipulation and data visualization.

See also

Author

Maintainer: Colin Fay contact@colinfay.me (ORCID)

Authors:

Sebastien Rochette sebastien@thinkr.fr (ORCID)

Other contributors:

ThinkR [copyright holder]